+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 
 ### Added
+- Set up testthat infrastructure and mock helpers (#7)
 - Implement extended Cristin endpoints (projects, keywords, funding) (#4)
 
 ### Fixed